cast abs() argument to int to fix compilation with BC++ 2007
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@47424 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/font.cpp
@@ -476,7 +476,8 @@ int wxNativeFontInfo::GetPointSize() const
     //        for printing!
     const int ppInch = ::GetDeviceCaps(ScreenHDC(), LOGPIXELSY);
 
-    return (int) (((72.0*(double)abs(lf.lfHeight)) / (double) ppInch) + 0.5);
+    // BC++ 2007 doesn't provide abs(long) overload, hence the cast
+    return (int) (((72.0*abs((int)lf.lfHeight)) / (double) ppInch) + 0.5);
 }
 
 wxSize wxNativeFontInfo::GetPixelSize() const
